ecommerce/requirements.txt
Django==2.2.11
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.0.0
djongo==1.3.1
pymongo==3.10.1
